+++ src/media_engine.cpp
@@ -15,6 +15,7 @@ extern "C" {
 #include <libavutil/channel_layout.h>
 #include <libavutil/error.h>
 #include <libavutil/opt.h>
+#include <libavutil/pixdesc.h>
 #include <libavutil/version.h>
 #include <libswresample/swresample.h>
 #include <libswscale/swscale.h>
@@ -1455,6 +1456,10 @@ class LibavMediaEngine final : public MediaEngine {
             info.channels = par->ch_layout.nb_channels;
             info.sample_rate = par->sample_rate;
             info.bit_depth = par->bits_per_raw_sample;
+            info.level = par->level > 0 ? par->level : 0;
+            if (par->codec_type == AVMEDIA_TYPE_VIDEO && par->color_trc != AVCOL_TRC_UNSPECIFIED)
+                if (const char* transfer = av_color_transfer_name(par->color_trc))
+                    info.color_transfer = transfer;
             if (par->bit_rate > 0) info.bitrate = static_cast<uint64_t>(par->bit_rate);
             info.default_stream = (stream->disposition & AV_DISPOSITION_DEFAULT) != 0;
             info.forced = (stream->disposition & AV_DISPOSITION_FORCED) != 0;